Refactor pybarsys to use `.env` files and add documentation
Enabler for #27
