+++ .env.example
@@ -1,0 +1,5 @@
+DATABASE_URL=sqlite:///db.sqlite3
+EMAIL_URL=smtp+tls://user:hunter2%40%3Axyz%3F%21@localhost:587
+LANGUAGE_CODE=en-us
+TIME_ZONE=Europe/Berlin
+SECRET_KEY=GENERATE THIS! See e.g. https://humberto.io/blog/tldr-generate-django-secret-key/
