+++ .travis.yml
@@ -3,4 +3,6 @@ python:
   - "3.6"
 install:
   - pip3 install -r requirements.txt
-script: python3 manage.py test
+  - cp .env.example .env
+script: 
+  - python3 manage.py test
