+++ requirements.txt
@@ -4,4 +4,5 @@ django-crispy-forms~=1.7.2
 django-debug-toolbar~=1.11
 django-filter~=2.1.0
 djangorestframework~=3.11.2
+django-environ~=0.4.5
 babel~=2.9.0
